refactor: https://github.com/metanorma/isodoc/issues/653
